requirements.txt
annotated-types==0.5.0
anyascii==0.3.2
argcomplete==3.1.1
astroid==2.15.6
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contractions==0.1.73
cymem==2.0.7
dill==0.3.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
flask==2.3.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathy==0.10.2
pickle5==0.0.11
platformdirs==3.10.0
preshed==3.0.8
pyahocorasick==2.0.0
pydantic==1.8.2
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn-pandas==2.2.0
smart-open==6.3.0
spacy==3.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
textblob==0.17.1
textsearch==0.0.24
thinc==8.0.17
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
typer==0.4.2
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.4
wasabi==0.10.1
werkzeug==2.3.7
wrapt==1.15.0
xmltodict==0.13.0
yq==3.2.2
zipp==3.16.2
